fleur: configs: thermal: Drop Xiaomi backlight brightness throttling
It does not work properly in AOSP/LineageOS and causes the backlight
to randomly change brightness out of nowhere. It is quite unclear to
me why this happens but since the brightness slider is not updated
with the values that mi_thermald sets, it's quite possible that we
are missing some MIUI framework change.

It also seems that we do not need to edit any other file which is
good since Xiaomi thermal configs are encrypted.

Signed-off-by: Elias Dimopoulos <linkdevel@protonmail.com>